% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj2graphml.R
\name{adj2graphml}
\alias{adj2graphml}
\title{Export an Adjacency Matrix to a graphml file}
\usage{
adj2graphml(adjMat, filename = "network.graphml", nodeAttributes,
  threshold = 0.5)
}
\arguments{
\item{adjMat}{A weighted adjacency matrix}

\item{filename}{Filename for the saved network.}

\item{nodeAttributes}{Data frame containing additional node information.}

\item{threshold}{Minimum threshold needed to create an edge.}
}
\value{
An igraph object of the input adjecency network.
}
\description{
Saves a weighted adjacency matrix to a graphML using a specified minimum threshold.
The resulting graphml file can be visualized with programs such as Cytoscape
}
\references{
https://github.com/iscb-dc-rsg/2016-summer-workshop/blob/master/3B-Hughitt-RNASeq-Coex-Network-Analysis/tutorial/README.Rmd
}

